Refactor DenseNet
