+++ hypothesis/nn/__init__.py
@@ -1,5 +1,3 @@
-r""""""
-
 # Multi Layered Perceptron
 from .multi_layered_perceptron import MultiLayeredPerceptron
 from .multi_layered_perceptron import MultiLayeredPerceptron as MLP
@@ -9,3 +7,7 @@ from .neuromodulation import NeuromodulatedELU
 from .neuromodulation import NeuromodulatedReLU
 from .neuromodulation import NeuromodulatedSELU
 from .neuromodulation import NeuromodulatedTanh
+# DenseNet
+from .densnet import DenseNetHead
+# ResNet
+from .resnet import ResNetHead
