+++ hypothesis/nn/densenet/__init__.py
@@ -1,0 +1,4 @@
+from .head import DenseNetHead
+from .network import DenseNet1D
+from .network import DenseNet2D
+from .network import DenseNet3D
